# src/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "# RNKM API\r\nThis is the documentation for https://freshersfairs.com",
        "title": "RNKM Backend",
        "contact": {
            "name": "ISD Team",
            "email": "sd.team.sgcu@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/baan": {
            "get": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Return the array of baan dto if successfully",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "baan"
                ],
                "summary": "Get all baans",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/proto.Baan"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseUnauthorizedErr"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseServiceDownErr"
                        }
                    }
                }
            }
        },
        "/baan/user": {
            "get": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Return the baan dto if successfully",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "baan"
                ],
                "summary": "Get the user's baan",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/proto.Baan"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseUnauthorizedErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseNotfoundErr"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseServiceDownErr"
                        }
                    }
                }
            }
        },
...
        "/estamp": {
            "get": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Get estamp id overview on what user has",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "summary": "Get user estamp",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/proto.GetUserEstampResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseUnauthorizedErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseInternalErr"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseServiceDownErr"
                        }
                    }
                }
            }
        },
...
        "/group": {
            "get": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Return the group dto if successfully",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "Get the group data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/proto.Group"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseUnauthorizedErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseNotfoundErr"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseServiceDownErr"
                        }
                    }
                }
            }
        },
...
        "/user": {
            "put": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Return the user dto if successfully",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create new user if it doesn't exist and Update the user data if exists",
                "parameters": [
                    {
                        "description": "user dto",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateUserDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/proto.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseBadRequestErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseUnauthorizedErr"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseServiceDownErr"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Return the user dto if successfully",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create new user",
                "parameters": [
                    {
                        "description": "User DTO",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/proto.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseBadRequestErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseUnauthorizedErr"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseForbiddenErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseNotfoundErr"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseServiceDownErr"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Return the user dto if successfully",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update the user data if exists",
                "parameters": [
                    {
                        "description": "user dto",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/proto.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseBadRequestErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseUnauthorizedErr"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseServiceDownErr"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.BadReqErrResponse": {
            "type": "object",
            "properties": {
                "failed_field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "value": {}
            }
        },
...
        "dto.ResponseBadRequestErr": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.BadReqErrResponse"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Invalid request body"
                },
                "status_code": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "dto.ResponseForbiddenErr": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Insufficiency permission"
                },
                "status_code": {
                    "type": "integer",
                    "example": 403
                }
            }
        },
...
        "dto.ResponseInternalErr": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Internal service error"
                },
                "status_code": {
                    "type": "integer",
                    "example": 500
                }
            }
        },
        "dto.ResponseNotfoundErr": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Not found"
                },
                "status_code": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
        "dto.ResponseServiceDownErr": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Service is down"
                },
                "status_code": {
                    "type": "integer",
                    "example": 503
                }
            }
        },
        "dto.ResponseUnauthorizedErr": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Invalid token"
                },
                "status_code": {
                    "type": "integer",
                    "example": 401
                }
            }
        },
...
        "dto.UpdateUserDto": {
            "type": "object",
            "properties": {
                "allergy_food": {
                    "type": "string"
                },
                "allergy_medicine": {
                    "type": "string"
                },
                "disease": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "food_restriction": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "line_id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.UserDto": {
            "type": "object",
            "required": [
                "can_select_baan",
                "firstname",
                "lastname",
                "line_id",
                "nickname",
                "phone",
                "title"
            ],
            "properties": {
                "allergy_food": {
                    "type": "string"
                },
                "allergy_medicine": {
                    "type": "string"
                },
                "can_select_baan": {
                    "type": "boolean"
                },
                "disease": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "food_restriction": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "line_id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "proto.Baan": {
            "type": "object",
            "properties": {
                "descriptionEN": {
                    "type": "string"
                },
                "descriptionTH": {
                    "type": "string"
                },
                "facebook": {
                    "type": "string"
                },
                "facebookUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "instagram": {
                    "type": "string"
                },
                "instagramUrl": {
                    "type": "string"
                },
                "line": {
                    "type": "string"
                },
                "lineUrl": {
                    "type": "string"
                },
                "nameEN": {
                    "type": "string"
                },
                "nameTH": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "proto.BaanInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "nameEN": {
                    "type": "string"
                },
                "nameTH": {
                    "type": "string"
                }
            }
        },
...
        "proto.Event": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "descriptionEN": {
                    "type": "string"
                },
                "descriptionTH": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nameEN": {
                    "type": "string"
                },
                "nameTH": {
                    "type": "string"
                }
            }
        },
...
        "proto.GetUserEstampResponse": {
            "type": "object",
            "properties": {
                "eventList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/proto.Event"
                    }
                }
            }
        },
        "proto.Group": {
            "type": "object",
            "properties": {
                "baans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/proto.BaanInfo"
                    }
                },
                "id": {
                    "type": "string"
                },
                "leaderID": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/proto.UserInfo"
                    }
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "proto.User": {
            "type": "object",
            "properties": {
                "allergyFood": {
                    "type": "string"
                },
                "allergyMedicine": {
                    "type": "string"
                },
                "baanId": {
                    "type": "string"
                },
                "canSelectBaan": {
                    "type": "boolean"
                },
                "disease": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "faculty": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "foodRestriction": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "isVerify": {
                    "type": "boolean"
                },
                "lastname": {
                    "type": "string"
                },
                "lineID": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "studentID": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "year": {
                    "type": "string"
                }
            }
        },
        "proto.UserInfo": {
            "type": "object",
            "properties": {
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AuthToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}